backend/requirements.txt
# ─── Core Django ───
Django>=5.1
djangorestframework>=3.15
django-cors-headers>=4.6
django-filter>=24.3
django-environ>=0.11
djangorestframework-simplejwt>=5.4

# ─── Database ───
# Using SQLite3 for local dev, Supabase (PostgreSQL) for production

# ─── AI / LLM ───
langchain>=0.3
langchain-google-genai>=2.0
langchain-groq>=0.2
langchain-community>=0.3
google-generativeai>=0.8
groq>=0.15

# ─── Web Scraping ───
beautifulsoup4>=4.12
requests>=2.32
lxml>=5.3

# ─── Media ───
Pillow>=11.1

# ─── Utilities ───
python-dotenv>=1.0
gunicorn>=23.0
whitenoise>=6.8
